Fix bugs with my last commit and edit some details.
public/css/main.css: Make the animation of the logo ease-in-out.
view/auth.signin: Make a pointer cursor on the Captcha.
view/contest.index: Fix the wrong href of the AC/Submission, add a paper-button in the <a>.
view/contest.list: Add a paper-button in the <a>.
view/contest.ranklist: Add a paper-button in the <a>.
view/errors.contest_end: Add header and footer on the page and make the whole page in the center.
view/layout.dashboard_nav: Do not stretch the picture
view/layout.header: Change the class of the dropdown menu to blue. Do not stretch the picture
view/problem.index: Modify the Font
view/problem.list: Add a paper-button in the <a>.
view/status.list: Add a paper-button in the <a>, fix most bugs of the Ajax.